refactor: refactored SeriesScreen split it to multiple sections.
